[not4land] spmma2 cuSPARSELt benchmarking
Summary:
Took
https://github.com/NVIDIA/CUDALibrarySamples/tree/master/cuSPARSELt/spmma2
and integrated it as an aten function in PyTorch for benchmarking
purposes. Made some modifications to the example (e.g., adding
matmulsearch) and some timing functions.

This PR is currently not intended to be landed. A future PR will
refractor the spmma2 example using custom class to expose the different
operations (pruning, compression, matmul) in python.

TODO: insert benchmark table here

ghstack-source-id: 5bba7ee66009ecc1c98164378092f2a560887ba9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/85835